refactor wizard configs
