+++ integrations/slack/definitions/index.ts
@@ -1,6 +1,5 @@
 export * from './actions'
 export * from './channels'
-export * from './configuration'
 export * from './events'
 export * from './secrets'
 export * from './states'
