+++ integrations/slack/src/oauth-wizard/index.ts
@@ -1,12 +1,10 @@
 import { generateRedirection } from '@botpress/common/src/html-dialogs'
 import { isOAuthWizardUrl, getInterstitialUrl } from '@botpress/common/src/oauth-wizard'
-import * as appManifestWizard from './appManifestWizard'
-import * as defaultConfigWizard from './defaultConfigWizard'
-import * as manualConfigWizard from './manualConfigWizard'
+import * as wizard from './wizard'
 import * as bp from '.botpress'
 
 export const oauthWizardHandler: bp.IntegrationProps['handler'] = async (props) => {
-  const { req, ctx, logger } = props
+  const { req, logger } = props
 
   if (!isOAuthWizardUrl(req.path)) {
     return {
@@ -16,13 +14,7 @@ export const oauthWizardHandler: bp.IntegrationProps['handler'] = async (props) 
   }
 
   try {
-    if (ctx.configurationType === 'refreshToken') {
-      return await manualConfigWizard.handler(props)
-    } else if (ctx.configurationType === 'manifestAppCredentials') {
-      return await appManifestWizard.handler(props)
-    } else {
-      return await defaultConfigWizard.handler(props)
-    }
+    return await wizard.handler(props)
   } catch (thrown: unknown) {
     const error = thrown instanceof Error ? thrown : Error(String(thrown))
     const errorMessage = 'OAuth wizard error: ' + error.message
