+++ integrations/slack/src/webhook-events/handlers/oauth-callback.ts
@@ -1,5 +1,6 @@
 import * as sdk from '@botpress/sdk'
 import { SlackClient } from 'src/slack-api'
+import { getAppCredentials } from 'src/slack-api/slack-manifest-client'
 import * as bp from '.botpress'
 
 export const isOAuthCallback = (req: sdk.Request): req is sdk.Request & { path: '/oauth' } => req.path === '/oauth'
@@ -14,10 +15,8 @@ export const handleOAuthCallback = async ({ req, client, ctx, logger }: bp.Handl
   }
 
   const slackClient = await SlackClient.createFromAuthorizationCode({ client, ctx, logger, authorizationCode: code })
-  const identifier =
-    ctx.configurationType === 'manifestAppCredentials' || ctx.configurationType === 'refreshToken'
-      ? slackClient.getBotUserId()
-      : slackClient.getTeamId()
+  const appCreds = await getAppCredentials(client, ctx)
+  const identifier = appCreds.clientId ? slackClient.getBotUserId() : slackClient.getTeamId()
 
   await client.configureIntegration({ identifier })
 }
